requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.0
black==24.4.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
crispy-bootstrap5==2024.2
cron-descriptor==1.4.3
Django==4.2.7
django-bootstrap5==24.1
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-ckeditor-5==0.2.12
django-crispy-forms==2.1
django-debug-toolbar==4.3.0
django-filter==24.2
django-jazzmin==3.0.0
django-modeltranslation==0.18.11
django-phonenumber-field==7.3.0
django-timezone-field==6.1.0
djangorestframework==3.15.1
factory-boy==3.3.0
Faker==24.11.0
flake8==7.0.0
flower==2.0.1
humanize==4.9.0
isort==5.13.2
kombu==5.3.7
Markdown==3.6
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
phonenumbers==8.13.35
pillow==10.3.0
platformdirs==4.2.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.2.0
python-crontab==3.0.0
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.4
six==1.16.0
sqlparse==0.5.0
tomli==2.0.1
tornado==6.4
typing_extensions==4.11.0
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
